update from class
