+++ wordle-web/src/router/index.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
-import Wordle from '../views/Wordle.vue'
+import Wordle from '../views/WordleView.vue'
 import GameView from '../views/GameView.vue'
 import WordleView from '../views/WordleView.vue'
